[4.6, 4.7, 4.8, trunk] Implement -Wwrong-language
The issue: In Android, many libraries share c and c++.
This can be a problem in such a library while building
with -Werror and using a flag that's only valid for c or c++.

An example of the warning:
cc1: warning: command line option '-std=gnu++11' is valid
for C++/ObjC++ but not for C [enabled by default]

The solution: -Wwrong-language has been implemented and
disabled by default. If enabled, -Wno-wrong-language works
as expected.